+++ Scribe/Components/UITableViewCells/AboutTableViewCell.swift
@@ -56,6 +56,16 @@ final class AboutTableViewCell: UITableViewCell {
       linkImageView.image = nil
     }
 
-    accessoryType = section.hasNestedNavigation ? .disclosureIndicator : .none
+    if section.hasNestedNavigation {
+      let disclosureIcon = UIImage(systemName: "chevron.right")
+      let accessory  = UIImageView(
+        frame: CGRect(
+          x: 0, y: 0, width: (disclosureIcon?.size.width)!, height: (disclosureIcon?.size.height)!
+        )
+      )
+      accessory.image = disclosureIcon
+      accessory.tintColor = menuOptionColor
+      accessoryView = accessory
+    }
   }
 }