+++ Scribe/InstallationTab/InstallationVC.swift
@@ -111,9 +111,6 @@ class InstallationVC: UIViewController {
       view.isUserInteractionEnabled = false
       view.backgroundColor = .clear
     }
-
-    topIconPhone.tintColor = .white
-    topIconPad.tintColor = .white
   }
 
   /// Sets properties for the app screen given the current device.
@@ -146,10 +143,6 @@ class InstallationVC: UIViewController {
 
     // Set link attributes for all textViews.
     for textView in allTextViews {
-      guard let linkBlueColor = UIColor(named: "linkBlue") else {
-        fatalError("Failed to load linkBlue color.")
-      }
-
       textView.linkTextAttributes = [
         NSAttributedString.Key.foregroundColor: linkBlueColor,
         NSAttributedString.Key.underlineStyle: NSUnderlineStyle.single.rawValue
@@ -162,12 +155,12 @@ class InstallationVC: UIViewController {
     let settingsSymbol: UIImage = getSettingsSymbol(fontSize: fontSize * 0.9)
     topIconPhone.image = settingsSymbol
     topIconPad.image = settingsSymbol
-    topIconPhone.tintColor = .init(.keyChar)
-    topIconPad.tintColor = .init(.keyChar)
+    topIconPhone.tintColor = UITraitCollection.current.userInterfaceStyle == .dark ? scribeCTAColor : keyCharColor
+    topIconPad.tintColor = UITraitCollection.current.userInterfaceStyle == .dark ? scribeCTAColor : keyCharColor
 
     // Enable installation directions and GitHub notice elements.
     settingsBtn.isUserInteractionEnabled = true
-    appTextBackground.backgroundColor = UIColor(named: "commandBar")
+    appTextBackground.backgroundColor = lightWhiteDarkBlackColor
 
     // Set the texts for the fields.
     switch Locale.userSystemLanguage {