+++ Scribe/Views/BaseTableViewController.swift
@@ -74,13 +74,17 @@ extension BaseTableViewController {
     if let reusableHeaderView = tableView.headerView(forSection: section) {
       headerView = reusableHeaderView
     } else {
-      headerView = UIView(frame: CGRect(x: 0, y: 0, width: tableView.bounds.width, height: sectionHeaderHeight))
+      headerView = UIView(
+        frame: CGRect(x: 0, y: 0, width: tableView.bounds.width, height: sectionHeaderHeight)
+      )
     }
 
-    let label = UILabel(frame: CGRect(x: 0, y: 0, width: headerView.bounds.width, height: sectionHeaderHeight))
+    let label = UILabel(
+      frame: CGRect(x: 0, y: 0, width: headerView.bounds.width, height: sectionHeaderHeight)
+    )
     label.text = dataSet[section].headingTitle
     label.font = UIFont.preferredFont(forTextStyle: .headline)
-    label.textColor = .black
+    label.textColor = keyCharColor
     headerView.addSubview(label)
 
     return headerView